split up into SetNewBudget and DisplayCurrentBudget
